Fase 6: pyproject, ruff, CI e glossario
- pyproject.toml con entry point `vedit`: dopo `pip install -e ".[dev]"` il
  comando si invoca senza `python -m`. La versione resta dichiarata solo in
  vedit/__init__.py e pyproject la legge da li'.
- ruff configurato per lint e formattazione; il codice e' pulito su E, W, F, I,
  UP, B, C4 e RUF. `ruff format` resta disponibile ma non obbligatorio, e il
  perche' e' scritto nel pyproject: i commenti allineati in colonna sui campi
  delle dataclass sono materiale di studio, e il formatter li schiaccerebbe.
- GitHub Action su push e pull request: installa ffmpeg, lancia lint e test su
  Python 3.10 e 3.12, poi genera il materiale di prova con ffmpeg e renderizza
  davvero il progetto demo verificando durata, dimensioni e che i fotogrammi
  non siano neri. Un export puo' uscire con codice 0 e produrre un video
  sbagliato: la CI lo controlla come lo controllerebbe una persona.
- docs/GLOSSARIO.md: i termini di montaggio usati nel repo, spiegati per chi
  viene dalla programmazione, con le trappole vere (i due significati di
  keyframe, perche' un crossfade accorcia il montaggio, pan/tilt che descrivono
  la camera e non l'immagine).

+++ video-toolkit/vedit/motion.py
@@ -96,7 +96,7 @@ def even_up(value: float) -> int:
     all'altro - si vede come un tremolio. Arrotondando per eccesso si evita anche
     di scendere sotto le dimensioni del canvas, cioe' di scoprire i bordi.
     """
-    ceiling = int(math.ceil(value - 1e-9))
+    ceiling = math.ceil(value - 1e-9)
     return ceiling + (ceiling % 2)
 
 
@@ -184,9 +184,9 @@ def _pan(clip, ctx: MotionContext, horizontal: bool, forward: bool):
         progress = _ramp(t, duration)
         offset = progress if forward else 1.0 - progress
         if horizontal:
-            x, y = int(round(travel_x * offset)), travel_y // 2
+            x, y = round(travel_x * offset), travel_y // 2
         else:
-            x, y = travel_x // 2, int(round(travel_y * offset))
+            x, y = travel_x // 2, round(travel_y * offset)
         return enlarged[y:y + canvas_h, x:x + canvas_w]
 
     # VideoClip ricava le dimensioni dal primo fotogramma: sono quelle del
